event_manager/registrations/serializers.py
from rest_framework import serializers
from .models import Registration


class RegistrationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Registration
        fields = ['user', 'event', 'registered_at']
        read_only_fields = ['registered_at']
