pyproject.toml
[tool.poetry]
name = "join-to-it-test-pet-project"
version = "0.1.0"
description = ""
authors = ["KittGT42 <114111458+KittGT42@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
setuptools = "^74.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
